Update comments and remove a package in the role tasks
Update the comment to correctly point at cisagov/cyhy-reports as the
source for the Python packages. Remove the python-pystache package from
the list as we have migrated to using chevron instead of pystache. I
was unable to find an apt package for chevron when checking the Debian
repositories so it will just be installed by pip when
cisagov/cyhy-reports is installed.