Update cyhy-reports installation logic
We now install the cyhy-reports Python package from the repository URL
instead of from the local filesystem. This not only follows our best
practices but is also required to clear idempotence on Debian Buster
hosts. This is because of a change in pip behavior where installing
from the filesystem forces a reinstall.

We still download the source blob and extract it locally because it is
still the most convenient way to get the two Python scripts and the
fonts onto the host.