Only configure the buf_size if explicitly provided
This changes the role's logic to only configure the buf_size for
texlive if the `texmf_buffer_size` variable is provided. This ensures
that the OS-provided configuration is used by default without this role
needing to keep track of the default value. In light of this change the
tests are reorganized to add a new scenario to verify functionality if
the variable is provided. The defaul scenario is modified to verify
that no change is made if it is not.