Update information about role variables
Add a note explaining the behavior behind optional variables with no
default values. This will hopefully make it clearer as to why they have
no default value and how they are used in this case. We cannot use the
`omit` special value because these variables are used in conditionals
for tasks instead of as module arguments.

Co-authored-by: Shane Frasier <jeremy.frasier@gwe.cisa.dhs.gov>