Add option to set ownership of created files
Add two variables that provide the ability to set ownership for any
non-system files and directories created by this role. The README is
updated accordingly. We specify non-system files because the texlive
configuration files added by this role have no reason to not be owned
by the root user (which is how they're explicitly configured).