Install a specific version of cisagov/cyhy-core
This adds a variable with a default value of the latest
cisagov/cyhy-core release's tag that is used to install the specified
version instead of whatever cisagov/ansible-role-cyhy-core defaults to
installing. This will ensure consistent deployments using this role as
we can specify the versions of both Python packages.

The value is also passed down to cisagov/ansible-role-ncats-webd since
it will _also_ install cisagov/cyhy-core using
cisagov/ansible-role-cyhy-core. This again ensures consistent
deployments when using this role.